move food and exercise package inside item package
