+++ src/main/java/seedu/duke/commands/Command.java
@@ -1,9 +1,8 @@
 package seedu.duke.commands;
 
-import seedu.duke.exercise.ExerciseList;
-import seedu.duke.food.FoodList;
+import seedu.duke.item.exercise.ExerciseList;
+import seedu.duke.item.food.FoodList;
 import seedu.duke.profile.Profile;
-import seedu.duke.ui.Statistics;
 import seedu.duke.ui.Ui;
 
 /**
