+++ src/main/java/seedu/duke/commands/DeleteExerciseCommand.java
@@ -1,7 +1,6 @@
 package seedu.duke.commands;
 
-import seedu.duke.exercise.Exercise;
-import seedu.duke.parser.Parser;
+import seedu.duke.item.exercise.Exercise;
 import seedu.duke.ui.Ui;
 
 import java.util.logging.Level;
