+++ src/main/java/seedu/duke/item/exercise/Exercise.java
@@ -1,4 +1,4 @@
-package seedu.duke.exercise;
+package seedu.duke.item.exercise;
 
 import seedu.duke.item.Item;
