+++ src/main/java/seedu/duke/storage/Decoder.java
@@ -1,9 +1,9 @@
 package seedu.duke.storage;
 
-import seedu.duke.exercise.Exercise;
-import seedu.duke.exercise.ExerciseList;
-import seedu.duke.food.Food;
-import seedu.duke.food.FoodList;
+import seedu.duke.item.exercise.Exercise;
+import seedu.duke.item.exercise.ExerciseList;
+import seedu.duke.item.food.Food;
+import seedu.duke.item.food.FoodList;
 import seedu.duke.profile.Profile;
 import seedu.duke.profile.exceptions.InvalidCharacteristicException;
